backend/app/routers/vision.py
from fastapi import APIRouter, UploadFile, File
from PIL import Image
import io

from typing import Annotated

router = APIRouter(prefix="/vision", tags=["vision"])

@router.post("/identify")
async def identify_item(file: Annotated[UploadFile, File(description="An image file to identify")]):
    """
    Identify an item from an uploaded image.

    UploadFile has the following attributes:
    - filename (str): The name of the uploaded file (e.g. myimage.jpg).
    - content_type (str): The content type of the file (MIME type / media type) (e.g. image/jpeg).
    - file (SpooledTemporaryFile): A "file-like" object that you can read() to get the bytes of the file.
    """
    image = Image.open(io.BytesIO(await file.read()))
    # TODO: Add ML model for image recognition
    # For now we return a placeholder response
    result = {
        "item": "apple",
        "confidence": 0.97
    }

    return result
