backend/pyproject.toml
[project]
name = "kura-api"
version = "0.1.0"
description = "FastAPI backend for Kura"
authors = [{ name = "Wilson Leung", email = "wilson.leung.2054@hotmail.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "pillow>=11.3.0",
    "pwdlib[argon2]>=0.2.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "ruff>=0.14.0",
    "torch>=2.8.0",
    "transformers>=4.57.0",
    "uvicorn>=0.37.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
